Deserialization handled.
Some comments: I'm not quite satisfied with the error handling, as
it is not the same everywhere in the code. parseTMachine returns
a Maybe TuringMachine but stringToAction uses `error`. Also in the
Turing module I use (!) to access the hashmap. Maybe I need to
refactor some things, but I'm kind of lazy.